refactor: (#9) 패키지 분리
