+++ src/infrastructure/domain/project/presentation/project.web.adapter.ts
@@ -1,10 +1,10 @@
 import { Body, Controller, Get, HttpCode, Post, Query } from '@nestjs/common';
-import { QueryCurrentRepositoryUseCase } from '../../../application/project/usecase/query-current-repository.usecase';
-import { CreateProjectUseCase } from '../../../application/project/usecase/create-project.usecase';
+import { QueryCurrentRepositoryUseCase } from '../../../../application/domain/project/usecase/query-current-repository.usecase';
+import { CreateProjectUseCase } from '../../../../application/domain/project/usecase/create-project.usecase';
 import { CreateProjectRequest } from './dto/project.web.dto';
-import { CurrentUser } from '../../global/decorator/current-user.decorator';
-import { User } from '../../../application/user/user';
-import { Permission } from '../../global/decorator/authority.decorator';
+import { CurrentUser } from '../../../global/decorator/current-user.decorator';
+import { User } from '../../../../application/domain/user/user';
+import { Permission } from '../../../global/decorator/authority.decorator';
 import { Authority } from '../../user/persistence/user.entity';
 
 @Controller('projects')
