+++ src/infrastructure/global/module/user.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UserTypeormEntity } from '../../user/persistence/user.entity';
-import { UserPort } from '../../../application/user/spi/user.spi';
-import { UserPersistenceAdapter } from '../../user/persistence/user.persistence.adapter';
+import { UserTypeormEntity } from '../../domain/user/persistence/user.entity';
+import { UserPort } from '../../../application/domain/user/spi/user.spi';
+import { UserPersistenceAdapter } from '../../domain/user/persistence/user.persistence.adapter';
 
 const USER_PORT = { provide: UserPort, useClass: UserPersistenceAdapter };
 const USER_REPOSITORY = TypeOrmModule.forFeature([UserTypeormEntity]);
